pyproject.toml
[project]
name = "music-platform-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12"
dependencies = [
    "cachetools>=5.5.2",
    "fastapi>=0.115.11",
    "librosa>=0.9.2",
    "matplotlib>=3.10.1",
    "numpy>=2.1.3",
    "opencv-python>=4.13.0.92",
    "pandas>=2.2.3",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.10.6",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tensorflow[and-cuda]>=2.19.0",
    "tqdm>=4.67.3",
    "uvicorn>=0.34.0",
]
